# perf: improves performance of ThumbnailPanel significantly