--- a/alt-tab-macos/ui/Application.swift
+++ b/alt-tab-macos/ui/Application.swift
@@ -98,6 +98,7 @@ class Application: NSApplication, NSApplicationDelegate, NSWindowDelegate {
     }
 
     func showUiOrCycleSelection(_ step: Int) {
+        NSLog("Application.showUiOrCycleSelection")
         appIsBeingUsed = true
         if isFirstSummon {
             isFirstSummon = false