@gingerr 2019-11-26
Btw. the first panel summon takes a little longer. Subsequent summons are improved greatly on my system.