@gingerr 2019-11-26
This only improves #45

Interesting documentation about layers: https://developer.apple.com/library/archive/documentation/Cocoa/Conceptual/CoreAnimation_guide/SettingUpLayerObjects/SettingUpLayerObjects.html