@lwouis 2019-11-26
I checked out the code, then removed the 2 layer lines and tested locally with like 50 finder windows: 1s to 1.1s. I then put back the 2 layer lines: 1s to 1.1s.

Basically I see no performance improvements with or without this PR