@gingerr 2019-11-26
Dammit. Made the same observation on my MacBook running the latest macOS, it makes no difference. My home computer is using an desktop cpu and running on nvidia drivers and also using an older macOS version (high sierra). No idea what of that is causing a difference there. Need to invest more time into this.

Thank you for checking it out and testing and profiling it. I also need to find a better lifecycle method (other than orderOut) cause visually it feels like the NSLog statement in orderOut() appears later in the console than the view actually disappearing. The console could be also running on a different and be a little off/async.