@lwouis 2019-11-26
Yes, basically I think we haven't really diagnosed the performance bottleneck correctly yet. First step should be to find a way to actually diagnose. I don't know why Instruments is not helping here :/