@gingerr 2019-11-26
Yeah instruments/profiler does not show the full story or i don't interpret it correctly and thats why i added the NSLog statements.

I did not check `AccessibilityApis` at all cause the expensive operation on my desktop system is something with images/layers. This class does is not related to fetching the images right? This is about identifying the real/relevant windows?