@lwouis 2019-11-26
> This class does is not related to fetching the images right? This is about identifying the real/relevant windows?

Calls to `AXUIElementCopyAttributeValue` are made to find if the window is of a minimum size (to avoid Chrome search box to show in the thumbnails since it's an `NSWindow` for some reason). For some reason these calls are expensive. Maybe they copy the image to measure the size at some point to get the dimensions? Who knows. The fact is that it's expensive so as much as possible we should avoid calling it. The same logic from the CG API family is cheap. Maybe checking the size there is enough? To be tested