@lwouis 2019-11-26
![image](https://user-images.githubusercontent.com/106195/69633770-32b2bc00-1094-11ea-85d3-120a3081217c.png)

Maybe the `Thread State Trace` tells the story? Maybe it is indeed the `WindowServer` taking time, and somehow that time doesn't appear in Instruments `Time Profiler`? I'm guessing the CG api to get the screenshots are XPC calls, so their execution happens in the `WindowServer` process instead of the app threads, so we don't see that time spent in the `Time Profiler`.