@koekeishiya 2019-11-26
@lwouis 

> Calls to AXUIElementCopyAttributeValue are made to find if the window is of a minimum size 

All AX API calls are expensive (and slow) as they are serialized requests. If one application takes a long time to respond, this will delay every other request even if you are doing them through multiple threads.

If all you care about is getting the current dimensions of the window, you can use the following private API call (located in SkyLight.framework).

```    
extern CGError SLSGetWindowBounds(int cid, uint32_t wid, CGRect *frame);

CGRect frame = {};
SLSGetWindowBounds(g_connection, window->id, &frame);
```