@lwouis 2019-12-26
I'm closing this PR as its improvements are unclear, and anyway I have significantly improved performance in an upcoming PR when I changed pretty much every OS call. See https://github.com/lwouis/alt-tab-macos/issues/14#issuecomment-569056863